requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1
# Instalamos pydantic con el extra [email] para garantizar email-validator
pydantic[email]==2.9.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
